final code
This app is inspired by an early app that was released on the App Store. Even though it's only function was showing a picture of a Red gem on the screen upon launch, it was priced at $999 and was created to "flaunt your wealth" and "brag" to others about purchasing it.